Moved project to subdirectory
