+++ .vscode/launch.json
@@ -10,7 +10,7 @@
             "request": "launch",
             "preLaunchTask": "build",
             // If you have changed target frameworks, make sure to update the program path.
-            "program": "${workspaceFolder}/bin/Debug/netcoreapp3.1/GloomhavenAbilityManager.dll",
+            "program": "${workspaceFolder}/Blazor/bin/Debug/netcoreapp3.1/GloomhavenAbilityManager.dll",
             "args": [],
             "cwd": "${workspaceFolder}",
             "stopAtEntry": false,
