+++ .vscode/tasks.json
@@ -7,7 +7,7 @@
             "type": "process",
             "args": [
                 "build",
-                "${workspaceFolder}/GloomhavenAbilityManager.csproj",
+                "${workspaceFolder}/Blazor/GloomhavenAbilityManager.Blazor.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -19,7 +19,7 @@
             "type": "process",
             "args": [
                 "publish",
-                "${workspaceFolder}/GloomhavenAbilityManager.csproj",
+                "${workspaceFolder}/Blazor/GloomhavenAbilityManager.Blazor.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -32,7 +32,7 @@
             "args": [
                 "watch",
                 "run",
-                "${workspaceFolder}/GloomhavenAbilityManager.csproj",
+                "${workspaceFolder}/Blazor/GloomhavenAbilityManager.Blazor.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
