+++ Blazor/Data/CharacterClassService.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace GloomhavenAbilityManager.Data
+namespace GloomhavenAbilityManager.Blazor.Data
 {
     public class CharacterClassService : ICharacterClassService
     {
