+++ Blazor/Data/ICharacterClassService.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace GloomhavenAbilityManager.Blazor.Data
+{
+    public interface ICharacterClassService
+    {
+        CharacterClass GetClass(int id);
+        Task<IEnumerable<CharacterClass>> GetClassesAsync();
+    }
+}
