+++ Blazor/Pages/Cards.razor
@@ -1,0 +1,56 @@
+﻿@page "/cards/{ClassIdParameter}"
+
+@using GloomhavenAbilityManager.Blazor.Data
+@inject IAbilityCardService CardService
+@inject ICharacterClassService ClassService
+
+<h1>Ability cards (@characterClass.Name)</h1>
+
+@if (cards == null)
+{
+    <p><em>Loading...</em></p>
+}
+else
+{
+    <table class="table">
+        <thead>
+            <tr>
+                <th>Class</th>
+                <th>Name</th>
+                <th>Level</th>
+            </tr>
+        </thead>
+        <tbody>
+            @foreach (var card in cards)
+            {
+            <tr>
+                <td><a href=@($"cards\\{card.ClassId}")>@ClassService.GetClass(card.ClassId).Name</a></td>
+                <td><a href=@($"card\\{card.Id}")>@card.Name</a></td>
+                <td>@card.Level</td>
+            </tr>
+            }
+        </tbody>
+    </table>
+}
+
+@code {
+    private IEnumerable<AbilityCardInfo> cards;
+    
+    [Parameter]
+    public string  ClassIdParameter {get; set; } = "All";
+
+    private int classId = 0;
+
+    private CharacterClass characterClass;
+
+    protected override async Task OnInitializedAsync()
+    {
+        if (! int.TryParse(ClassIdParameter, out classId))
+        {
+            classId = 0;
+        }
+
+        cards = await CardService.GetCardsAsync(classId);
+        characterClass = ClassService.GetClass(classId);
+    }
+}
