+++ Blazor/Pages/Char.razor
@@ -1,6 +1,6 @@
 @page "/char/{CharacterIdParameter}"
 
-@using GloomhavenAbilityManager.Data
+@using GloomhavenAbilityManager.Blazor.Data
 @inject IAbilityCardService CardService
 @inject ICharacterClassService ClassService
 @inject ICharacterService CharacterService
