+++ Blazor/Pages/SelectableCard.razor
@@ -1,4 +1,4 @@
-@using GloomhavenAbilityManager.Data
+@using GloomhavenAbilityManager.Blazor.Data
 @inject IAbilityCardService CardService
 @inject ICharacterClassService ClassService
