+++ Blazor/_Imports.razor
@@ -5,5 +5,5 @@
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.JSInterop
-@using GloomhavenAbilityManager
-@using GloomhavenAbilityManager.Shared
+@using GloomhavenAbilityManager.Blazor
+@using GloomhavenAbilityManager.Blazor.Shared
