+++ GloomhavenAbilityManager.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29613.14
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GloomhavenAbilityManager", "GloomhavenAbilityManager.csproj", "{7E7C1C19-B23C-4CC7-8538-93620067060C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GloomhavenAbilityManager", "Blazor\GloomhavenAbilityManager.Blazor.csproj", "{7E7C1C19-B23C-4CC7-8538-93620067060C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
