--- Pages/Card.razor
@@ -1,33 +1,0 @@
-@page "/card/{CardIdParameter}"
-
-@using GloomhavenAbilityManager.Data
-@inject IAbilityCardService CardService
-@inject ICharacterClassService ClassService
-
-@if (card == null)
-{
-    <p><em>Loading...</em></p>
-}
-else
-{
-    <p>@card.Name (@ClassService.GetClass(card.ClassId).Name @card.Level)</p>
-    <img src="@card.ImagePath" height="320" width="240">
-}
-
-@code {
-    
-    [Parameter]
-    public string CardIdParameter {get; set; } = "0";
-
-    private AbilityCardInfo card;
-
-    protected override async Task OnInitializedAsync()
-    {
-        if ( ! int.TryParse(CardIdParameter, out int _id))
-        {
-            _id = 0;
-        }
-
-        card = await CardService.GetCardAsync(_id);
-    }
-}
