--- Shared/NavMenu.razor
@@ -1,46 +1,0 @@
-﻿@using GloomhavenAbilityManager.Data
-@inject ICharacterService CharacterService
-
-<div class="top-row pl-4 navbar navbar-dark">
-    <a class="navbar-brand" href="">Gloomhaven</a>
-    <button class="navbar-toggler" @onclick="ToggleNavMenu">
-        <span class="navbar-toggler-icon"></span>
-    </button>
-</div>
-
-<div class="@NavMenuCssClass" @onclick="ToggleNavMenu">
-    <ul class="nav flex-column">
-        <li class="nav-item px-3">
-            <NavLink class="nav-link" href="" Match="NavLinkMatch.All">
-                <span class="oi oi-home" aria-hidden="true"></span> Home
-            </NavLink>
-        </li>
-
-        @foreach (var character in characters)
-        {
-            <li class="nav-item px-3">
-                <NavLink class="nav-link" href=@($"char\\{character.Id}")>
-                    <span class="oi oi-list-rich" aria-hidden="true"></span> @character.Name
-                </NavLink>
-            </li>
-        }
-    </ul>
-</div>
-
-@code {
-    private bool collapseNavMenu = true;
-
-    private IEnumerable<Character> characters;
-
-    private string NavMenuCssClass => collapseNavMenu ? "collapse" : null;
-
-    protected override async Task OnInitializedAsync()
-    {
-        characters = await CharacterService.GetCharactersAsync();
-    }
-
-    private void ToggleNavMenu()
-    {
-        collapseNavMenu = !collapseNavMenu;
-    }
-}
